// apps/web/src/api/generated/model/attendanceResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { AttendanceEmployeeDto } from './attendanceEmployeeDto';
import type { AttendanceResponseDtoLunchDutyType } from './attendanceResponseDtoLunchDutyType';

export interface AttendanceResponseDto {
  username: string;
  date: string;
  /** @nullable */
  morningCheckin: string | null;
  /** @nullable */
  morningCheckout: string | null;
  /** @nullable */
  noonCheck: string | null;
  /** @nullable */
  afternoonCheckin: string | null;
  /** @nullable */
  afternoonCheckout: string | null;
  /** @nullable */
  morningCheckinImage: string | null;
  /** @nullable */
  morningCheckoutImage: string | null;
  /** @nullable */
  noonCheckImage: string | null;
  /** @nullable */
  afternoonCheckinImage: string | null;
  /** @nullable */
  afternoonCheckoutImage: string | null;
  /** @nullable */
  note: string | null;
  /**
     * Lunch duty type
     * @nullable
     */
  lunchDutyType?: AttendanceResponseDtoLunchDutyType;
  employee?: AttendanceEmployeeDto;
  createdAt: string;
  updatedAt: string;
}