// apps/web/src/api/generated/model/attendanceCommandControllerCheckAttendanceFromWebBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BTN HRMS API
 * The Human Resource Management System API description
 * OpenAPI spec version: 1.0
 */
import type { AttendanceCommandControllerCheckAttendanceFromWebBodyImageSource } from './attendanceCommandControllerCheckAttendanceFromWebBodyImageSource';
import type { AttendanceCommandControllerCheckAttendanceFromWebBodyLunchDutyType } from './attendanceCommandControllerCheckAttendanceFromWebBodyLunchDutyType';
import type { AttendanceCommandControllerCheckAttendanceFromWebBodySession } from './attendanceCommandControllerCheckAttendanceFromWebBodySession';
import type { AttendanceCommandControllerCheckAttendanceFromWebBodyType } from './attendanceCommandControllerCheckAttendanceFromWebBodyType';

export type AttendanceCommandControllerCheckAttendanceFromWebBody = {
  date: string;
  session: AttendanceCommandControllerCheckAttendanceFromWebBodySession;
  type: AttendanceCommandControllerCheckAttendanceFromWebBodyType;
  note?: string;
  latitude?: string;
  longitude?: string;
  location?: string;
  imageSource?: AttendanceCommandControllerCheckAttendanceFromWebBodyImageSource;
  lunchDutyType?: AttendanceCommandControllerCheckAttendanceFromWebBodyLunchDutyType;
  image?: Blob | File;
};